A.I. Code
This submissions contains the final code. Populations contain chromosomes made up of genes which determine an angle and associated number of steps. When the population is propagated, a chromosome which is likely 'weak' (high cost value), will 'die' and another will be born from two fit chromosomes -inheriting half the genes from both parents. Randomness is added in the form of chance mutations in genes or for random genes to be dropped or added. A graph production api is included which shows costs for chromosomes over generations.